+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 - Email: Don't set a default template per account by default. Only a default for the user.
+- Email: Fix download all attachments as zip
 
 21-02-2022 6.6.44
 - pr2analyzer: fix CSV report